+++ src/hydra_umc_bridge_droids/__init__.py
@@ -9,6 +9,7 @@
 from hydra_umc_sdk.bridge_contract import BridgeJob, CellState, JobPhase, MachineState
 
 from .coordinator import DroidActionPlan, DroidCoordinator, DroidDispatch
+from .platform_profiles import MOBILE_MANIPULATOR, MOBILE_ONLY, PROFILES, PlatformProfile, SimulatedDroid, check_against_profile
 from .spot_transport import SpotDroidControl, SpotSendResult, open_bosdyn_robot_command
 
 __all__ = [
@@ -19,6 +20,12 @@ __all__ = [
     "DroidActionPlan",
     "DroidCoordinator",
     "DroidDispatch",
+    "PlatformProfile",
+    "SimulatedDroid",
+    "MOBILE_ONLY",
+    "MOBILE_MANIPULATOR",
+    "PROFILES",
+    "check_against_profile",
     "SpotDroidControl",
     "SpotSendResult",
     "open_bosdyn_robot_command",